beginning of refactor
